Fixed a few little offset things.
Weird thing I noticed is that I though we were storing pointers in the Index_Node.  However, a pointer is 8 bytes, which would mean only 32 pointers would
fit in an indirect block, exactly half of the 64 we need.  Thus, you are right on storing block number, but it is really weird.